Drop new `Account` and associated APIs
I moved them to an internal struct in the sqlite crate. I think I'd like to bring it back eventually, but with a more focused and intentional PR.
